Add telemetry fixture requirement
